Set Board2D.LENGTH as int for compatibility

+++ app/src/main/java/chess/space/environment/Board2D.java
@@ -59,7 +59,7 @@ public class Board2D implements Serializable {
     }
 
 
-    public long getLength() {
+    public int getLength() {
         return LENGTH;
     }
